Switched to hex_simd and made wallet RPC calls more reliable
Firstly, switched from the hex library to hex_simd, so we can get that speeeeed the server is gonna need.

Secondly, RPC calls constantly failing was really damn annoying, so I made it to where the server will attempt to send a request up to 6 times before actually returning an error. While this obviously won't completely eliminate reqwest errors, it's been pretty reliable so far (from keeping the server up for an hour and waiting for an RPC call to fail).

Since RPC calls are so so so slow, I'm working on trying to cache them as much as possible to try to avoid having to send them at all. As an example, instead of sending a daemon call request it every time. Then, if for whatever reason the daemon call fails on the first initialization attempt, we just use an old known fee amount instead. I'm planning on working to have way more RPC calls cached and updated over time, in order to improve the overall performance of the server.