Type checking for DB + Fundraiser searching
To avoid the annoying issue of accidentally writing to the wrong database, I made it to where you need to use the typed versions of the databases (like AUthDB) that are just wrappers around the regular trees.

Also added the ability to search through the fundraisers and have different ways of retrieving said data.