Withdrawing + typed db + better reliability
This is a really huge commit lol. This commit was originally just supposed to be the ability to withdraw Monero to your account. However, while implementing this, I ran into numerous annoying issues with both writing the code for it as well as the actual monero service itself, which I'll talk about now.

So firstly, when the sled Tree type is incredibly annoying, though not at the fault of spacejam. Since every singel Tree looks exactly the same, if you screw up when passing inputs to a function (for example, in warp), you can very easily start writing data to the wrong Tree, ruining the database and requiring a lovely database cleansing. In order to solve this, I created the db.rs file, and wrote a bunch of generic functions so that, basically, all of the de/serialization is done completely automatically, and the compiler actually enforces type checking rules. This is also great because it reduces the boiler plate of serializing and deserializing every single value every time I want to read/write from the db. The main downside is that I can't reuse certain serializations (such as the username, which is often passed to multiple functions), but honestly, it's worth the minuscule performance downgrade to not have to deal with those bugs.

Secondly, when I finished typed databases and actually got the code for the withdrawal going, I noticed something very bad. So for whatever reason, sometimes when sending a Monero transaction to the DB, the transaction would go through, but because the Monero wallet RPC likes taking it's sweet time responding, the entire wallet's balance would go down but the user's balance would remain the same. This is badddddd news, because users could just keep requesting a withdrawal and drain the wallet. Now there are two solutions to this that the Monero wallet itself has. One is called Monero accounts, in which a new full Monero address is generated for each new user. This has the advantage of being able to keep user's finances completely separate from one another, so that if someone manages to take over an entire account, they would only have access to their funds. The main downside is that we'd have to either send two monero transfers whenever someone withdraws (in order to snatch up the service fee), or we'd have to have a more complicated way of managing accounts. It's still a huge advantage over just sending everything to the main wallet with payment id's, though, but there's an even better solution.

Monero has another thing called subaddresses, which are similar to accounts but they still give the money to the account that controls the subaddress. By using subaddresses instead of accounts, we get all the nice advantages of accounts like having individual monero balances and higher security, but we also lose the only real disadvantage of acounts. Not onyl that, but many monero wallet RPC calls actually support querying subindexes of an account (so specific subaddresses), which means that we're actually making fewer RPC calls overall. We still do need to keep track of the amount of service fees taken per subaddress, but that's easy enough to do with he db (just add a service_fee to the struct). The struct service fee isn't even necessary, it's just a small precaution just in case I have buggy code somewhere, so once again the entire financial database is completely fault tolerant thanks to the Monero blockchain.

I also fixed some bugs where, for example, users were given their balance instead of their unlocked (confirmed by the Monero blockchain) balance. This would obviously be really bad if a malicious user to were to send malicious transactions that our node picked up on for even a moment. I also fixed some RPC typo bugs.