Fixed get_balance and high CPU usage
I'm legitimately so stupid it hurts.

Firstly, I fixed a bug with the get_balance function where it literally wouldn't work if there was more than one user, because I wasn't inserting a comma in the array I generated :(.

Secondly, I reduced the idle CPU usage from 200% to between 0.1 and 1% on release builds by adding a sleep call to a function.