Switched from Argon 2 to Blake 3 + Logging
So firstly, I decided to switch to blake 3 for hashing, and have the actual argon2 passord has be done on the client side. This should help take some stress off of the server.

Secondly, I implemented logging and inserted several (very long jesus) logging statements throughout the code. Every time a log is sent, it's stored in a 65 KB text buffer. Every 90 seconds, or whenever the text buffer hits 65 kb, it's appended to a file. The actual logs themselves are stored in a JSON structure.

I thought of splitting this commit into two different ones, but I didn't care enough to tbh. No one else reads these but me lol :(