Software uploads + Captcha fails faster
We now have the beginnings of being able to upload software to the server! Whenever a user wants to, we set up a web socket connection (we use websockets since we still want to be web compatible), and then we have a custom binary protocol to upload stuff to the webserver. Now the actual protocol itself is a little complicated, but basically, we compress all of the files using either zstd or brotli, and then split each of those compressed files into 64KiB chunks. First, some information about how many chunks there is is sent, as well as the information about the files that are going to be sent and their hashes. All of this information MUST be extremely accurate, since if it's off by even a single byte, the server will (eventually) just reject everything sent by that client. Then, we set up  a stack allocated buffer of 64 KiB, and begin receiving data about the chunks into said buffer. We use 64 KiB since the maximum packet size is 8KiB, so it's just a convenient rounding up. Every time the chunk buffer is filled (or we have finished receiving all the chunks), we do an async file write to a cache for storing chunks.
Once all the chunks have been sent, we *should* just save it into some directory for easy access, and call it a day. What we currently do is decompress it into the root directory of the game repository, which is obviously extremely dangerous. DO NOT RUN THIS WEBSOCKET SERVER PUBLICLY. I CAN GUARANTEE THERE ARE MANY EASY WAYS OF EITHER CRASHING THE SERVER OR WRITING TO COMPLETELY ARBITRARY FILES ON THE SYSTEM.

Oh and I made the captcha check happen write at the start of logging in, just to save some CPU cycles in the event of an invalid captcha.