Added searching
Software search is currently very very simple. It just checks to see if the search term is in the title, and then it just pushes that item (as a search result) into a Vec. It only starts actually doing sorting work if the Vec has reached capacity (the amount of items requested). When that happens, it finds the worst item in the Vec (by just iterating over it with a linear search) and returns a mutable reference to said item. Once it does that, it compares the worst item with the item next in the queue, and if the item in the queue is better, the worst item is replaced with it.

Now there are definetely optimizations that can be made. For one, we could literally just use par_iter_mut and instantly get a speed boost (once we have a large number of results, with a low number it'll probably actually be slower than iter_mut). For another, we could try sorting it constantly, so that the search is quicker. Who knows, we should definetely benchmark benchmark benchmark and find what's fastest.

There are only two search algorithms implemented at the moment, so that's definetely something to be worked on lol. It may also be advantageous to cache very popular searches short term (like the Trending algorithm with a blank search), so we should work on that.