Check if archives are valid before adding software
Basically the title. I just decompress the file (writing to /dev/null,
very smart I know) and if Zstd throws an error while decompressing, I
just panic the tokio task and kill it. Once again, should probably send
an error and just break the connection in a clean and pretty way, but
this is easier for now.

Fixes #6