Added images to software uploads
Now software is required to have an image that goes alongside it. The image is a compressed ZSTD file that is stored as the has of the file. The front-end simply needs to query the view_software_image API and download the image from there. Hopefully then it will cache the image (since hashes are stored, it literally will never ever need to query the server for that image again).

I also switched back to stable since RUst 1.60 supports incremental compilation again, which I need for development.