Switched from using SystemTime to using OffsetDateTime
So SystemTime is great, but it has some issues. The main one I was worried about when making the switch is timezones. If for whatever reason we had data being transferred from two different timeszones, we moved the servers from one timezone to another, etc, it would mean that all of our data would require a movement to whatever the new timezone is. This can be a frustratinga nd annoying future bug, so I decided to just start using the time library and getting UTC only. We can still use Unix timestamps which is nice.